+++ packages/docs/content/docs/index.mdx
@@ -19,7 +19,7 @@ In a nutshell, here are the most interesting features offered:
 | Type-checked         | Option requirements    | Paragraphs and lists    |
 | Highly configurable  | Value constraints      | Custom error phrases    |
 | Browser-compatible   | Name suggestions       | [SGR] colors and styles |
-| 34KB minified        | Subcommands            | Option grouping/hiding  |
+| 35KB minified        | Subcommands            | Option grouping/hiding  |
 | ESM-native           | Asynchronous callbacks | Usage statements        |
 | Online documentation | GNU's _short-options_  |                         |
