+++ packages/docs/content/docs/library/formatter.mdx
@@ -142,7 +142,7 @@ A **text** section can be used to write various kinds of content, such as introd
 
 > [!TIP]
 >
-> You can create a personalized footer including the repository URL through the `sectionFooter` function, which can be imported from `'tsargp/utility'{:ts}`.
+> You can create a personalized footer including the repository URL through the `sectionFooter` function.
 
 ### Usage section
