+++ packages/docs/content/docs/library/options.mdx
@@ -536,7 +536,7 @@ The `version` attribute, if present, specifies the version information. Its valu
 
 > [!TIP]
 >
-> You can get the `version` field from a JSON module through the `getVersion` function, which can be imported from `'tsargp/utility'{:ts}`.
+> You can get the `version` field from a JSON module through the `getVersion` function.
 
 ### Command option
