+++ packages/docs/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tsargp/docs",
-  "version": "1.11.0",
+  "version": "1.11.1",
   "description": "Documentation for tsargp",
   "homepage": "https://dsogari.github.io/tsargp/docs",
   "license": "MIT",
