+++ packages/tsargp/package.json
@@ -18,7 +18,7 @@
   },
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
+      "types": "./dist/library/index.d.ts",
       "default": "./dist/index.min.js"
     },
     "./enums": {
@@ -34,8 +34,8 @@
       "default": "./dist/library/index.js"
     },
     "./utility": {
-      "types": "./dist/utility/index.d.ts",
-      "default": "./dist/utility/index.js"
+      "types": "./dist/library/utils.d.ts",
+      "default": "./dist/library/utils.js"
     },
     "./package": "./package.json"
   },
@@ -58,9 +58,9 @@
   "license": "MIT",
   "scripts": {
     "prepare": "bun run build",
-    "build": "tsc -b && tsc -b src && bun build dist/index.js --outfile dist/index.min.js --minify && sed -i 's/const enum/enum/g' dist/enums/*.d.ts"
+    "build": "tsc -b && tsc -b src && bun build dist/library/index.js --outfile dist/index.min.js --minify && sed -i 's/const enum/enum/g' dist/enums/*.d.ts"
   },
   "type": "module",
   "main": "./dist/index.min.js",
-  "types": "./dist/index.d.ts"
+  "types": "./dist/library/index.d.ts"
 }
