--- packages/tsargp/src/index.ts
@@ -1,33 +1,0 @@
-export type * from './library/index.js';
-export {
-  config,
-  allHelpItems,
-  envHelpItems,
-  ErrorItem,
-  HelpItem,
-  cs,
-  tf,
-  fg,
-  bg,
-  ul,
-  parse,
-  parseInto,
-  validate,
-  format,
-  allOf,
-  oneOf,
-  /** @deprecated use `not` instead */
-  not as notOf,
-  not,
-  valuesFor,
-  seq,
-  style,
-  ext8,
-  rgb,
-  AnsiString,
-  AnsiMessage,
-  ErrorMessage,
-  JsonMessage,
-  TextMessage,
-  WarnMessage,
-} from './library/index.js';
