+++ packages/tsargp/src/library/index.ts
@@ -1,4 +1,3 @@
-export { allOf, not, oneOf, valuesFor } from './common.js';
 export * from './config.js';
 export * from './enums.js';
 export * from './formatter.js';
@@ -6,3 +5,4 @@ export * from './options.js';
 export * from './parser.js';
 export * from './styles.js';
 export * from './validator.js';
+export { allOf, not, oneOf, valuesFor, getVersion, numberInRange, sectionFooter } from './utils.js';
