--- packages/tsargp/src/utility/index.ts
@@ -1,2 +1,0 @@
-export * from '../library/common.js';
-export * from '../library/utils.js';
